requirements.txt
fastapi
uvicorn[standard]
httpx
playwright
pydantic
pandas
numpy
openpyxl
pypdf
pdfplumber
matplotlib
google-genai
python-dotenv
beautifulsoup4
lxml